# openapi/minimovie-api.yaml
openapi: 3.0.3
info:
  title: MiniMovie API
  description: API for retrieving intuitive movie, TV series, and person information
  version: 1.0.0
servers:
  - url: https://api.minimovie.info
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /auth/{provider}:
    get:
      summary: Begin OAuth login
      description: >-
        Redirects the user to the chosen OAuth provider's authorization page.
      operationId: beginAuth
      security: []
      parameters:
        - name: provider
          in: path
          required: true
          description: OAuth provider name
          schema:
            type: string
            enum:
              - google
              - apple
      responses:
        "302":
          description: Redirect to OAuth provider
        "400":
          description: Unknown provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/token:
    post:
      summary: Exchange auth code for session token
      description: >-
        Exchanges a one-time auth code (received from the OAuth callback redirect)
        for a raw session token. Requires the guest API key as a Bearer token.
      operationId: exchangeToken
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TokenRequest"
      responses:
        "200":
          description: Session token returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          description: Invalid request body or missing code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized (invalid guest key)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "410":
          description: Code expired or already exchanged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/session:
    get:
      summary: Get current session
      description: Returns the authenticated user's profile and unseen achievement count
      operationId: getSession
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Session info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me/delete:
    post:
      summary: Delete account
      description: >-
        Permanently deletes the authenticated user's account and all associated data.
        Requires a confirmation field set to "delete".
      operationId: deleteAccount
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeleteAccountRequest"
      responses:
        "200":
          description: Account deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteAccountResponse"
        "400":
          description: Missing or incorrect confirmation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/export:
    post:
      summary: Export user data
      description: >-
        Exports the authenticated user's data as a JSON file download.
        Rate-limited to one export per day.
      operationId: exportUserData
      security:
        - sessionCookie: []
      responses:
        "200":
          description: JSON file download
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="minimovie-data-export.json"'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExportData"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Export limit reached (1 per day)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users/me/achievements:
    get:
      summary: List achievements
      description: Returns all achievements earned by the authenticated user
      operationId: getAchievements
      security:
        - sessionCookie: []
      responses:
        "200":
          description: User achievements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AchievementsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/achievements/unseen:
    get:
      summary: List unseen achievements
      description: Returns achievements not yet seen by the user
      operationId: getUnseenAchievements
      security:
        - sessionCookie: []
      responses:
        "200":
          description: Unseen achievements
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AchievementsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/me/achievements/seen:
    patch:
      summary: Mark achievements as seen
      description: Marks all unseen achievements as seen for the authenticated user
      operationId: markAchievementsSeen
      security:
        - sessionCookie: []
      responses:
        "204":
          description: Achievements marked as seen
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Guest API key used for public catalog routes and token exchange
    sessionCookie:
      type: apiKey
      in: cookie
      name: mm_session
      description: Session cookie set after OAuth login (production name is __Host-mm_session)
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: HTTP status text
        message:
          type: string
          description: Detailed error message
...
    TokenRequest:
      type: object
      required:
        - code
      properties:
        code:
          type: string
          description: One-time auth code from the OAuth callback redirect
    TokenResponse:
      type: object
      properties:
        session_token:
          type: string
          description: Raw session token to be stored as a cookie
    SessionResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/SessionUser"
        unseenAchievementCount:
          type: integer
    SessionUser:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
          nullable: true
        givenName:
          type: string
          nullable: true
          description: User's first name from the OAuth provider
        avatarUrl:
          type: string
          nullable: true
          description: Profile picture URL from the OAuth provider (Google only)
    DeleteAccountRequest:
      type: object
      required:
        - confirm
      properties:
        confirm:
          type: string
          description: Must be set to "delete" to confirm
          example: delete
    DeleteAccountResponse:
      type: object
      properties:
        deleted:
          type: boolean
          example: true
    ExportData:
      type: object
      properties:
        account:
          type: object
          properties:
            username:
              type: string
              nullable: true
            givenName:
              type: string
              nullable: true
            avatarUrl:
              type: string
              nullable: true
            providers:
              type: array
              items:
                type: string
            createdAt:
              type: string
              format: date-time
        watchlist:
          type: array
          items:
            $ref: "#/components/schemas/WatchlistItem"
        watchEvents:
          type: array
          items:
            $ref: "#/components/schemas/WatchEvent"
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/UserAchievement"
    WatchlistItem:
      type: object
      properties:
        id:
          type: string
        mediaType:
          type: string
          enum:
            - movie
            - series
        mediaId:
          type: integer
        mediaTitle:
          type: string
        posterPath:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - want_to_watch
            - watched
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        lastWatchedAt:
          type: string
          format: date-time
          nullable: true
        watchCount:
          type: integer
        genres:
          type: array
          items:
            type: string
        runtimeMinutes:
          type: integer
          nullable: true
        voteAverage:
          type: number
          format: float
          nullable: true
        releaseYear:
          type: integer
          nullable: true
        addedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    WatchEvent:
      type: object
      properties:
        id:
          type: string
        mediaType:
          type: string
          enum:
            - movie
            - episode
            - series
            - season
        mediaId:
          type: integer
        mediaTitle:
          type: string
        seriesId:
          type: integer
          nullable: true
        seriesTitle:
          type: string
          nullable: true
        seasonNumber:
          type: integer
          nullable: true
        episodeNumber:
          type: integer
          nullable: true
        episodeCount:
          type: integer
          nullable: true
          description: Total episodes in the series/season (for series/season-level events)
        seasonCount:
          type: integer
          nullable: true
          description: Total seasons in the series (for series-level events)
        watchedAt:
          type: string
          format: date-time
          nullable: true
        timezone:
          type: string
        rewatchNumber:
          type: integer
        runtimeMinutes:
          type: integer
          nullable: true
        genres:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    UserAchievement:
      type: object
      properties:
        id:
          type: string
        achievementId:
          type: string
        name:
          type: string
          description: Human-readable achievement name
        description:
          type: string
          description: Short explanation of how the achievement is earned
        earnedViaMediaType:
          type: string
          nullable: true
          enum:
            - movie
            - series
          description: The media type of the item that triggered the achievement
        earnedViaMediaId:
          type: integer
          nullable: true
          description: The media ID of the item that triggered the achievement
        earnedViaMediaTitle:
          type: string
          nullable: true
          description: The title of the media that triggered the achievement
        seenAt:
          type: string
          format: date-time
          nullable: true
        earnedAt:
          type: string
          format: date-time
    AchievementsResponse:
      type: object
      properties:
        achievements:
          type: array
          items:
            $ref: "#/components/schemas/UserAchievement"